+++ R/LearnerClassifQDA.R
@@ -26,9 +26,12 @@ LearnerClassifQDA = R6Class("LearnerClassifQDA", inherit = LearnerClassif,
   public = list(
     initialize = function() {
       ps = ParamSet$new(list(
+        ParamUty$new("prior", tags = "train"),
         ParamFct$new("method", default = "moment", levels = c("moment", "mle", "mve", "t"), tags = "train"),
+        ParamInt$new("nu", tags = "train"),
         ParamFct$new("predict.method", default = "plug-in", levels = c("plug-in", "predictive", "debiased", "looCV"), tags = "predict")
       ))
+      ps$add_dep("nu", "method", CondEqual$new("t"))
 
       super$initialize(
         id = "classif.qda",