+++ R/LearnerClassifSVM.R
@@ -26,19 +26,25 @@ LearnerClassifSVM = R6Class("LearnerClassifSVM", inherit = LearnerClassif,
     initialize = function() {
       ps = ParamSet$new(list(
         ParamFct$new("type", default = "C-classification", levels = c("C-classification", "nu-classification"), tags = "train"),
-        ParamDbl$new("cost", default = 1, lower = 0, tags = "train"), # requires = quote(type == "C-classification")),
-        ParamDbl$new("nu", default = 0.5, tags = "train"), # requires = quote(type == "nu-classification")),
+        ParamDbl$new("cost", default = 1, lower = 0, tags = "train"),
+        ParamDbl$new("nu", default = 0.5, tags = "train"),
         ParamFct$new("kernel", default = "radial", levels = c("linear", "polynomial", "radial", "sigmoid"), tags = "train"),
-        ParamInt$new("degree", default = 3L, lower = 1L, tags = "train"), # requires = quote(kernel == "polynomial")),
-        ParamDbl$new("coef0", default = 0, tags = "train"), # requires = quote(kernel == "polynomial" || kernel == "sigmoid")),
-        ParamDbl$new("gamma", lower = 0, tags = "train"), # requires = quote(kernel != "linear")),
+        ParamInt$new("degree", default = 3L, lower = 1L, tags = "train"),
+        ParamDbl$new("coef0", default = 0, tags = "train"),
+        ParamDbl$new("gamma", lower = 0, tags = "train"),
         ParamDbl$new("cachesize", default = 40L, tags = "train"),
         ParamDbl$new("tolerance", default = 0.001, lower = 0, tags = "train"),
         ParamLgl$new("shrinking", default = TRUE, tags = "train"),
         ParamInt$new("cross", default = 0L, lower = 0L, tags = "train"), # tunable = FALSE),
         ParamLgl$new("fitted", default = TRUE, tags = "train"), # tunable = FALSE),
-        ParamUty$new("scale", default = TRUE, tags = "train") # , tunable = TRUE)
+        ParamUty$new("scale", default = TRUE, tags = "train"), # , tunable = TRUE)
+        ParamUty$new("class.weights", default = NULL, tags = "train")
       ))
+      ps$add_dep("cost", "type", CondEqual$new("C-classification"))
+      ps$add_dep("nu", "type", CondEqual$new("nu-classification"))
+      ps$add_dep("degree", "kernel", CondEqual$new("polynomial"))
+      ps$add_dep("coef0", "kernel", CondAnyOf$new(c("polynomial", "sigmoid")))
+      ps$add_dep("gamma", "kernel", CondAnyOf$new(c("polynomial", "radial", "sigmoid")))
 
       super$initialize(
         id = "classif.svm",