+++ NEWS.md
@@ -1,3 +1,9 @@
+# mlr3learners 0.1.5
+
+* Added parameter and parameter dependencies to `regr.glmnet`, `regr.km`, `regr.ranger`, `regr.svm`, `regr.xgboost`, `classif.glmnet`, `classif.lda`, `classif.naivebayes`, `classif.qda`, `classif.ranger` and `classif.svm`.
+* glmnet: Added `relax` parameter (v3.0)
+* xgboost: Updated parameters to v0.90.0.2
+
 # mlr3learners 0.1.4
 
 * Fixed a bug in `*.xgboost` and `*.svm` which was triggered if columns